refactor(core): migrate to hexagonal architecture and decouple mcp server logic
