+++ Dockerfile
@@ -23,8 +23,10 @@ RUN pip install --no-cache-dir --upgrade pip
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copy the rest of the application code
-COPY main.py .
-# COPY session_string_generator.py . # Optional: if needed within the container, otherwise can be run outside
+COPY pyproject.toml README.md ./
+COPY src/ src/
+
+RUN pip install --no-cache-dir .
 
 # Create a non-root user and switch to it
 RUN adduser --disabled-password --gecos "" appuser && chown -R appuser:appuser /app
@@ -44,4 +46,4 @@ ENV TELEGRAM_SESSION_STRING=""
 # EXPOSE 8000
 
 # Define the command to run the application
-CMD ["python", "main.py"] 
+CMD ["telegram-mcp"]
