+++ src/telegram_mcp/chats.py
@@ -1,0 +1,890 @@
+import os
+import telethon
+import time
+import asyncio
+from typing import List, Dict, Optional, Union, Any
+from pathlib import Path
+from telethon.tl.types import *
+from telethon.tl.functions.messages import GetForumTopicsRequest, ReadDiscussionRequest
+from telethon import functions, types, utils
+from mcp.server.fastmcp import Context
+
+# Use absolute imports for our own module as it's the standard for PyPI packages
+from telegram_mcp.client import client, logger
+from telegram_mcp.utils import *
+from telegram_mcp.security import *
+
+
+async def get_chats(page: int = 1, page_size: int = 20) -> str:
+    """
+    Get a paginated list of chats.
+    Args:
+        page: Page number (1-indexed).
+        page_size: Number of chats per page.
+    """
+    try:
+        dialogs = await client.get_dialogs()
+        start = (page - 1) * page_size
+        end = start + page_size
+        if start >= len(dialogs):
+            return "Page out of range."
+        chats = dialogs[start:end]
+        lines = []
+        for dialog in chats:
+            entity = dialog.entity
+            chat_id = entity.id
+            title = getattr(entity, "title", None) or getattr(entity, "first_name", "Unknown")
+            lines.append(f"Chat ID: {chat_id}, Title: {title}")
+        return "\n".join(lines)
+    except Exception as e:
+        return log_and_format_error("get_chats", e)
+
+
+async def list_topics(
+    chat_id: int,
+    limit: int = 200,
+    offset_topic: int = 0,
+    search_query: str = None,
+) -> str:
+    """
+    Retrieve forum topics from a supergroup with the forum feature enabled.
+
+    Note for LLM: You can send a message to a selected topic via reply_to_message tool
+    by using Topic ID as the message_id parameter.
+
+    Args:
+        chat_id: The ID of the forum-enabled chat (supergroup).
+        limit: Maximum number of topics to retrieve.
+        offset_topic: Topic ID offset for pagination.
+        search_query: Optional query to filter topics by title.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+
+        if not isinstance(entity, Channel) or not getattr(entity, "megagroup", False):
+            return "The specified chat is not a supergroup."
+
+        if not getattr(entity, "forum", False):
+            return "The specified supergroup does not have forum topics enabled."
+
+        result = await client(
+            functions.channels.GetForumTopicsRequest(
+                channel=entity,
+                offset_date=0,
+                offset_id=0,
+                offset_topic=offset_topic,
+                limit=limit,
+                q=search_query or None,
+            )
+        )
+
+        topics = getattr(result, "topics", None) or []
+        if not topics:
+            return "No topics found for this chat."
+
+        messages_map = {}
+        if getattr(result, "messages", None):
+            messages_map = {message.id: message for message in result.messages}
+
+        lines = []
+        for topic in topics:
+            line_parts = [f"Topic ID: {topic.id}"]
+
+            title = getattr(topic, "title", None) or "(no title)"
+            line_parts.append(f"Title: {title}")
+
+            total_messages = getattr(topic, "total_messages", None)
+            if total_messages is not None:
+                line_parts.append(f"Messages: {total_messages}")
+
+            unread_count = getattr(topic, "unread_count", None)
+            if unread_count:
+                line_parts.append(f"Unread: {unread_count}")
+
+            if getattr(topic, "closed", False):
+                line_parts.append("Closed: Yes")
+
+            if getattr(topic, "hidden", False):
+                line_parts.append("Hidden: Yes")
+
+            top_message_id = getattr(topic, "top_message", None)
+            top_message = messages_map.get(top_message_id)
+            if top_message and getattr(top_message, "date", None):
+                line_parts.append(f"Last Activity: {top_message.date.isoformat()}")
+
+            lines.append(" | ".join(line_parts))
+
+        return "\n".join(lines)
+    except Exception as e:
+        return log_and_format_error(
+            "list_topics",
+            e,
+            chat_id=chat_id,
+            limit=limit,
+            offset_topic=offset_topic,
+            search_query=search_query,
+        )
+
+
+async def list_chats(
+    chat_type: str = None, limit: int = 20, unread_only: bool = False, unmuted_only: bool = False
+) -> str:
+    """
+    List available chats with metadata.
+
+    Args:
+        chat_type: Filter by chat type ('user', 'group', 'channel', or None for all)
+        limit: Maximum number of chats to retrieve from Telegram API (applied before filtering, so fewer results may be returned when filters are active).
+        unread_only: If True, only return chats with unread messages.
+        unmuted_only: If True, only return unmuted chats.
+    """
+    try:
+        dialogs = await client.get_dialogs(limit=limit)
+
+        results = []
+        for dialog in dialogs:
+            entity = dialog.entity
+
+            # Filter by type if requested
+            current_type = get_entity_filter_type(entity)
+
+            if chat_type and current_type != chat_type.lower():
+                continue
+
+            # Format chat info
+            chat_info = f"Chat ID: {entity.id}"
+
+            if hasattr(entity, "title"):
+                chat_info += f", Title: {entity.title}"
+            elif hasattr(entity, "first_name"):
+                name = f"{entity.first_name}"
+                if hasattr(entity, "last_name") and entity.last_name:
+                    name += f" {entity.last_name}"
+                chat_info += f", Name: {name}"
+
+            chat_info += f", Type: {get_entity_type(entity)}"
+
+            if hasattr(entity, "username") and entity.username:
+                chat_info += f", Username: @{entity.username}"
+
+            # Add unread count if available
+            unread_count = getattr(dialog, "unread_count", 0) or 0
+            # Also check unread_mark (manual "mark as unread" flag)
+            inner_dialog = getattr(dialog, "dialog", None)
+            unread_mark = (
+                bool(getattr(inner_dialog, "unread_mark", False)) if inner_dialog else False
+            )
+
+            # Extract mute status from notify_settings
+            notify_settings = getattr(inner_dialog, "notify_settings", None)
+            mute_until = getattr(notify_settings, "mute_until", None)
+            if mute_until is None:
+                is_muted = False
+            elif isinstance(mute_until, datetime):
+                is_muted = mute_until.timestamp() > time.time()
+            else:
+                is_muted = mute_until > time.time()
+
+            # Filter by mute status if requested
+            if unmuted_only and is_muted:
+                continue
+
+            # Filter by unread status if requested
+            if unread_only and unread_count == 0 and not unread_mark:
+                continue
+
+            if unread_count > 0:
+                chat_info += f", Unread: {unread_count}"
+            elif unread_mark:
+                chat_info += ", Unread: marked"
+            else:
+                chat_info += ", No unread messages"
+
+            chat_info += f", Muted: {'yes' if is_muted else 'no'}"
+
+            # Add unread mentions count if available
+            unread_mentions = getattr(dialog, "unread_mentions_count", 0) or 0
+            if unread_mentions > 0:
+                chat_info += f", Unread mentions: {unread_mentions}"
+
+            results.append(chat_info)
+
+        if not results:
+            return f"No chats found matching the criteria."
+
+        return "\n".join(results)
+    except Exception as e:
+        return log_and_format_error(
+            "list_chats",
+            e,
+            chat_type=chat_type,
+            limit=limit,
+            unread_only=unread_only,
+            unmuted_only=unmuted_only,
+        )
+
+
+async def get_chat(chat_id: Union[int, str]) -> str:
+    """
+    Get detailed information about a specific chat.
+
+    Args:
+        chat_id: The ID or username of the chat.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+
+        result = []
+        result.append(f"ID: {entity.id}")
+
+        is_user = isinstance(entity, User)
+
+        if hasattr(entity, "title"):
+            result.append(f"Title: {entity.title}")
+            result.append(f"Type: {get_entity_type(entity)}")
+            if hasattr(entity, "username") and entity.username:
+                result.append(f"Username: @{entity.username}")
+
+            # Fetch participants count reliably
+            try:
+                participants_count = (await client.get_participants(entity, limit=0)).total
+                result.append(f"Participants: {participants_count}")
+            except Exception as pe:
+                result.append(f"Participants: Error fetching ({pe})")
+
+        elif is_user:
+            name = f"{entity.first_name}"
+            if entity.last_name:
+                name += f" {entity.last_name}"
+            result.append(f"Name: {name}")
+            result.append(f"Type: {get_entity_type(entity)}")
+            if entity.username:
+                result.append(f"Username: @{entity.username}")
+            if entity.phone:
+                result.append(f"Phone: {entity.phone}")
+            result.append(f"Bot: {'Yes' if entity.bot else 'No'}")
+            result.append(f"Verified: {'Yes' if entity.verified else 'No'}")
+
+        # Get last activity if it's a dialog
+        try:
+            # Using get_dialogs might be slow if there are many dialogs
+            # Alternative: Get entity again via get_dialogs if needed for unread count
+            dialog = await client.get_dialogs(limit=1, offset_id=0, offset_peer=entity)
+            if dialog:
+                dialog = dialog[0]
+                result.append(f"Unread Messages: {dialog.unread_count}")
+                if dialog.message:
+                    last_msg = dialog.message
+                    sender_name = "Unknown"
+                    if last_msg.sender:
+                        sender_name = getattr(last_msg.sender, "first_name", "") or getattr(
+                            last_msg.sender, "title", "Unknown"
+                        )
+                        if hasattr(last_msg.sender, "last_name") and last_msg.sender.last_name:
+                            sender_name += f" {last_msg.sender.last_name}"
+                    sender_name = sender_name.strip() or "Unknown"
+                    result.append(f"Last Message: From {sender_name} at {last_msg.date}")
+                    result.append(f"Message: {last_msg.message or '[Media/No text]'}")
+        except Exception as diag_ex:
+            logger.warning(f"Could not get dialog info for {chat_id}: {diag_ex}")
+            pass
+
+        return "\n".join(result)
+    except Exception as e:
+        return log_and_format_error("get_chat", e, chat_id=chat_id)
+
+
+async def get_admins(chat_id: Union[int, str]) -> str:
+    """
+    Get all admins in a group or channel.
+    """
+    try:
+        # Fix: Use the correct filter type ChannelParticipantsAdmins
+        participants = await client.get_participants(chat_id, filter=ChannelParticipantsAdmins())
+        lines = [
+            f"ID: {p.id}, Name: {getattr(p, 'first_name', '')} {getattr(p, 'last_name', '')}".strip()
+            for p in participants
+        ]
+        return "\n".join(lines) if lines else "No admins found."
+    except Exception as e:
+        logger.exception(f"get_admins failed (chat_id={chat_id})")
+        return log_and_format_error("get_admins", e, chat_id=chat_id)
+
+
+async def subscribe_public_channel(channel: Union[int, str]) -> str:
+    """
+    Subscribe (join) to a public channel or supergroup by username or ID.
+    """
+    try:
+        entity = await resolve_entity(channel)
+        await client(functions.channels.JoinChannelRequest(channel=entity))
+        title = getattr(entity, "title", getattr(entity, "username", "Unknown channel"))
+        return f"Subscribed to {title}."
+    except telethon.errors.rpcerrorlist.UserAlreadyParticipantError:
+        title = getattr(entity, "title", getattr(entity, "username", "this channel"))
+        return f"Already subscribed to {title}."
+    except telethon.errors.rpcerrorlist.ChannelPrivateError:
+        return "Cannot subscribe: this channel is private or requires an invite link."
+    except Exception as e:
+        return log_and_format_error("subscribe_public_channel", e, channel=channel)
+
+
+async def get_direct_chat_by_contact(contact_query: str) -> str:
+    """
+    Find a direct chat with a specific contact by name, username, or phone.
+
+    Args:
+        contact_query: Name, username, or phone number to search for.
+    """
+    try:
+        result = await client(functions.contacts.GetContactsRequest(hash=0))
+        contacts = result.users
+        found_contacts = []
+        for contact in contacts:
+            if not contact:
+                continue
+            name = (
+                f"{getattr(contact, 'first_name', '')} {getattr(contact, 'last_name', '')}".strip()
+            )
+            username = getattr(contact, "username", "")
+            phone = getattr(contact, "phone", "")
+            if (
+                contact_query.lower() in name.lower()
+                or (username and contact_query.lower() in username.lower())
+                or (phone and contact_query in phone)
+            ):
+                found_contacts.append(contact)
+        if not found_contacts:
+            return f"No contacts found matching '{contact_query}'."
+        results = []
+        dialogs = await client.get_dialogs()
+        for contact in found_contacts:
+            contact_name = (
+                f"{getattr(contact, 'first_name', '')} {getattr(contact, 'last_name', '')}".strip()
+            )
+            for dialog in dialogs:
+                if isinstance(dialog.entity, User) and dialog.entity.id == contact.id:
+                    chat_info = f"Chat ID: {dialog.entity.id}, Contact: {contact_name}"
+                    if getattr(contact, "username", ""):
+                        chat_info += f", Username: @{contact.username}"
+                    if dialog.unread_count:
+                        chat_info += f", Unread: {dialog.unread_count}"
+                    results.append(chat_info)
+                    break
+        if not results:
+            found_names = ", ".join(
+                [f"{c.first_name} {c.last_name}".strip() for c in found_contacts]
+            )
+            return f"Found contacts: {found_names}, but no direct chats were found with them."
+        return "\n".join(results)
+    except Exception as e:
+        return log_and_format_error("get_direct_chat_by_contact", e, contact_query=contact_query)
+
+
+async def get_contact_chats(contact_id: Union[int, str]) -> str:
+    """
+    List all chats involving a specific contact.
+
+    Args:
+        contact_id: The ID or username of the contact.
+    """
+    try:
+        contact = await resolve_entity(contact_id)
+        if not isinstance(contact, User):
+            return f"ID {contact_id} is not a user/contact."
+        contact_name = (
+            f"{getattr(contact, 'first_name', '')} {getattr(contact, 'last_name', '')}".strip()
+        )
+        direct_chat = None
+        dialogs = await client.get_dialogs()
+        results = []
+        for dialog in dialogs:
+            if isinstance(dialog.entity, User) and dialog.entity.id == contact_id:
+                chat_info = f"Direct Chat ID: {dialog.entity.id}, Type: Private"
+                if dialog.unread_count:
+                    chat_info += f", Unread: {dialog.unread_count}"
+                results.append(chat_info)
+                break
+        common_chats = []
+        try:
+            common = await client.get_common_chats(contact)
+            for chat in common:
+                chat_type = get_entity_type(chat)
+                chat_info = f"Chat ID: {chat.id}, Title: {chat.title}, Type: {chat_type}"
+                results.append(chat_info)
+        except:
+            results.append("Could not retrieve common groups.")
+        if not results:
+            return f"No chats found with {contact_name} (ID: {contact_id})."
+        return f"Chats with {contact_name} (ID: {contact_id}):\n" + "\n".join(results)
+    except Exception as e:
+        return log_and_format_error("get_contact_chats", e, contact_id=contact_id)
+
+
+async def get_last_interaction(contact_id: Union[int, str]) -> str:
+    """
+    Get the most recent message with a contact.
+
+    Args:
+        contact_id: The ID or username of the contact.
+    """
+    try:
+        contact = await resolve_entity(contact_id)
+        if not isinstance(contact, User):
+            return f"ID {contact_id} is not a user/contact."
+        contact_name = (
+            f"{getattr(contact, 'first_name', '')} {getattr(contact, 'last_name', '')}".strip()
+        )
+        messages = await client.get_messages(contact, limit=5)
+        if not messages:
+            return f"No messages found with {contact_name} (ID: {contact_id})."
+        results = [f"Last interactions with {contact_name} (ID: {contact_id}):"]
+        for msg in messages:
+            sender = "You" if msg.out else contact_name
+            message_text = msg.message or "[Media/No text]"
+            results.append(f"Date: {msg.date}, From: {sender}, Message: {message_text}")
+        return "\n".join(results)
+    except Exception as e:
+        return log_and_format_error("get_last_interaction", e, contact_id=contact_id)
+
+
+async def get_message_context(
+    chat_id: Union[int, str], message_id: int, context_size: int = 3
+) -> str:
+    """
+    Retrieve context around a specific message.
+
+    Args:
+        chat_id: The ID or username of the chat.
+        message_id: The ID of the central message.
+        context_size: Number of messages before and after to include.
+    """
+    try:
+        chat = await resolve_entity(chat_id)
+        messages_before = await client.get_messages(chat, limit=context_size, max_id=message_id)
+        central_message = await client.get_messages(chat, ids=message_id)
+        if central_message is not None and (not isinstance(central_message, list)):
+            central_message = [central_message]
+        elif central_message is None:
+            central_message = []
+        messages_after = await client.get_messages(
+            chat, limit=context_size, min_id=message_id, reverse=True
+        )
+        if not central_message:
+            return f"Message with ID {message_id} not found in chat {chat_id}."
+        all_messages = list(messages_before) + list(central_message) + list(messages_after)
+        all_messages.sort(key=lambda m: m.id)
+        results = [f"Context for message {message_id} in chat {chat_id}:"]
+        for msg in all_messages:
+            sender_name = get_sender_name(msg)
+            highlight = " [THIS MESSAGE]" if msg.id == message_id else ""
+            reply_content = ""
+            if msg.reply_to and msg.reply_to.reply_to_msg_id:
+                try:
+                    replied_msg = await client.get_messages(chat, ids=msg.reply_to.reply_to_msg_id)
+                    if replied_msg:
+                        replied_sender = "Unknown"
+                        if replied_msg.sender:
+                            replied_sender = getattr(
+                                replied_msg.sender, "first_name", ""
+                            ) or getattr(replied_msg.sender, "title", "Unknown")
+                        reply_content = f" | reply to {msg.reply_to.reply_to_msg_id}\n  → Replied message: [{replied_sender}] {replied_msg.message or '[Media/No text]'}"
+                except Exception:
+                    reply_content = (
+                        f" | reply to {msg.reply_to.reply_to_msg_id} (original message not found)"
+                    )
+            results.append(
+                f"ID: {msg.id} | {sender_name} | {msg.date}{highlight}{reply_content}\n{msg.message or '[Media/No text]'}\n"
+            )
+        return "\n".join(results)
+    except Exception as e:
+        return log_and_format_error(
+            "get_message_context",
+            e,
+            chat_id=chat_id,
+            message_id=message_id,
+            context_size=context_size,
+        )
+
+
+async def create_group(title: str, user_ids: List[Union[int, str]]) -> str:
+    """
+    Create a new group or supergroup and add users.
+
+    Args:
+        title: Title for the new group
+        user_ids: List of user IDs or usernames to add to the group
+    """
+    try:
+        users = []
+        for user_id in user_ids:
+            try:
+                user = await resolve_entity(user_id)
+                users.append(user)
+            except Exception as e:
+                logger.error(f"Failed to get entity for user ID {user_id}: {e}")
+                return f"Error: Could not find user with ID {user_id}"
+        if not users:
+            return "Error: No valid users provided"
+        try:
+            result = await client(functions.messages.CreateChatRequest(users=users, title=title))
+            if hasattr(result, "chats") and result.chats:
+                created_chat = result.chats[0]
+                return f"Group created with ID: {created_chat.id}"
+            elif hasattr(result, "chat") and result.chat:
+                return f"Group created with ID: {result.chat.id}"
+            elif hasattr(result, "chat_id"):
+                return f"Group created with ID: {result.chat_id}"
+            else:
+                await asyncio.sleep(1)
+                dialogs = await client.get_dialogs(limit=5)
+                for dialog in dialogs:
+                    if dialog.title == title:
+                        return f"Group created with ID: {dialog.id}"
+                return f"Group created successfully. Please check your recent chats for '{title}'."
+        except Exception as create_err:
+            if "PEER_FLOOD" in str(create_err):
+                return "Error: Cannot create group due to Telegram limits. Try again later."
+            else:
+                raise
+    except Exception as e:
+        logger.exception(f"create_group failed (title={title}, user_ids={user_ids})")
+        return log_and_format_error("create_group", e, title=title, user_ids=user_ids)
+
+
+async def invite_to_group(group_id: Union[int, str], user_ids: List[Union[int, str]]) -> str:
+    """
+    Invite users to a group or channel.
+
+    Args:
+        group_id: The ID or username of the group/channel.
+        user_ids: List of user IDs or usernames to invite.
+    """
+    try:
+        entity = await resolve_entity(group_id)
+        users_to_add = []
+        for user_id in user_ids:
+            try:
+                user = await resolve_entity(user_id)
+                users_to_add.append(user)
+            except ValueError as e:
+                return f"Error: User with ID {user_id} could not be found. {e}"
+        try:
+            result = await client(
+                functions.channels.InviteToChannelRequest(channel=entity, users=users_to_add)
+            )
+            invited_count = 0
+            if hasattr(result, "users") and result.users:
+                invited_count = len(result.users)
+            elif hasattr(result, "count"):
+                invited_count = result.count
+            return f"Successfully invited {invited_count} users to {entity.title}"
+        except telethon.errors.rpcerrorlist.UserNotMutualContactError:
+            return "Error: Cannot invite users who are not mutual contacts. Please ensure the users are in your contacts and have added you back."
+        except telethon.errors.rpcerrorlist.UserPrivacyRestrictedError:
+            return (
+                "Error: One or more users have privacy settings that prevent you from adding them."
+            )
+        except Exception as e:
+            return log_and_format_error("invite_to_group", e, group_id=group_id, user_ids=user_ids)
+    except Exception as e:
+        logger.error(
+            f"telegram_mcp invite_to_group failed (group_id={group_id}, user_ids={user_ids})",
+            exc_info=True,
+        )
+        return log_and_format_error("invite_to_group", e, group_id=group_id, user_ids=user_ids)
+
+
+async def leave_chat(chat_id: Union[int, str]) -> str:
+    """
+    Leave a group or channel by chat ID.
+
+    Args:
+        chat_id: The chat ID or username to leave.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+        if isinstance(entity, Channel):
+            try:
+                await client(functions.channels.LeaveChannelRequest(channel=entity))
+                chat_name = getattr(entity, "title", str(chat_id))
+                return f"Left channel/supergroup {chat_name} (ID: {chat_id})."
+            except Exception as chan_err:
+                return log_and_format_error("leave_chat", chan_err, chat_id=chat_id)
+        elif isinstance(entity, Chat):
+            try:
+                me = await client.get_me(input_peer=True)
+                await client(
+                    functions.messages.DeleteChatUserRequest(chat_id=entity.id, user_id=me)
+                )
+                chat_name = getattr(entity, "title", str(chat_id))
+                return f"Left basic group {chat_name} (ID: {chat_id})."
+            except Exception as chat_err:
+                logger.warning(
+                    f"First leave attempt failed: {chat_err}, trying alternative method"
+                )
+                try:
+                    me_full = await client.get_me()
+                    await client(
+                        functions.messages.DeleteChatUserRequest(
+                            chat_id=entity.id, user_id=me_full.id
+                        )
+                    )
+                    chat_name = getattr(entity, "title", str(chat_id))
+                    return f"Left basic group {chat_name} (ID: {chat_id})."
+                except Exception as alt_err:
+                    return log_and_format_error("leave_chat", alt_err, chat_id=chat_id)
+        else:
+            entity_type = type(entity).__name__
+            return log_and_format_error(
+                "leave_chat",
+                Exception(
+                    f"Cannot leave chat ID {chat_id} of type {entity_type}. This function is for groups and channels only."
+                ),
+                chat_id=chat_id,
+            )
+    except Exception as e:
+        logger.exception(f"leave_chat failed (chat_id={chat_id})")
+        error_str = str(e).lower()
+        if "invalid" in error_str and "chat" in error_str:
+            return log_and_format_error(
+                "leave_chat",
+                Exception(
+                    f"Error leaving chat: This appears to be a channel/supergroup. Please check the chat ID and try again."
+                ),
+                chat_id=chat_id,
+            )
+        return log_and_format_error("leave_chat", e, chat_id=chat_id)
+
+
+async def get_participants(chat_id: Union[int, str]) -> str:
+    """
+    List all participants in a group or channel.
+    Args:
+        chat_id: The group or channel ID or username.
+    """
+    try:
+        participants = await client.get_participants(chat_id)
+        lines = [
+            f"ID: {p.id}, Name: {getattr(p, 'first_name', '')} {getattr(p, 'last_name', '')}"
+            for p in participants
+        ]
+        return "\n".join(lines)
+    except Exception as e:
+        return log_and_format_error("get_participants", e, chat_id=chat_id)
+
+
+async def create_channel(title: str, about: str = "", megagroup: bool = False) -> str:
+    """
+    Create a new channel or supergroup.
+    """
+    try:
+        result = await client(
+            functions.channels.CreateChannelRequest(title=title, about=about, megagroup=megagroup)
+        )
+        return f"Channel '{title}' created with ID: {result.chats[0].id}"
+    except Exception as e:
+        return log_and_format_error(
+            "create_channel", e, title=title, about=about, megagroup=megagroup
+        )
+
+
+async def edit_chat_title(chat_id: Union[int, str], title: str) -> str:
+    """
+    Edit the title of a chat, group, or channel.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+        if isinstance(entity, Channel):
+            await client(functions.channels.EditTitleRequest(channel=entity, title=title))
+        elif isinstance(entity, Chat):
+            await client(functions.messages.EditChatTitleRequest(chat_id=chat_id, title=title))
+        else:
+            return f"Cannot edit title for this entity type ({type(entity)})."
+        return f"Chat {chat_id} title updated to '{title}'."
+    except Exception as e:
+        logger.exception(f"edit_chat_title failed (chat_id={chat_id}, title='{title}')")
+        return log_and_format_error("edit_chat_title", e, chat_id=chat_id, title=title)
+
+
+async def delete_chat_photo(chat_id: Union[int, str]) -> str:
+    """
+    Delete the photo of a chat, group, or channel.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+        if isinstance(entity, Channel):
+            await client(
+                functions.channels.EditPhotoRequest(channel=entity, photo=InputChatPhotoEmpty())
+            )
+        elif isinstance(entity, Chat):
+            await client(
+                functions.messages.EditChatPhotoRequest(
+                    chat_id=chat_id, photo=InputChatPhotoEmpty()
+                )
+            )
+        else:
+            return f"Cannot delete photo for this entity type ({type(entity)})."
+        return f"Chat {chat_id} photo deleted."
+    except Exception as e:
+        logger.exception(f"delete_chat_photo failed (chat_id={chat_id})")
+        return log_and_format_error("delete_chat_photo", e, chat_id=chat_id)
+
+
+async def get_banned_users(chat_id: Union[int, str]) -> str:
+    """
+    Get all banned users in a group or channel.
+    """
+    try:
+        participants = await client.get_participants(
+            chat_id, filter=ChannelParticipantsKicked(q="")
+        )
+        lines = [
+            f"ID: {p.id}, Name: {getattr(p, 'first_name', '')} {getattr(p, 'last_name', '')}".strip()
+            for p in participants
+        ]
+        return "\n".join(lines) if lines else "No banned users found."
+    except Exception as e:
+        logger.exception(f"get_banned_users failed (chat_id={chat_id})")
+        return log_and_format_error("get_banned_users", e, chat_id=chat_id)
+
+
+async def get_invite_link(chat_id: Union[int, str]) -> str:
+    """
+    Get the invite link for a group or channel.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+        try:
+            from telethon.tl import functions
+
+            result = await client(functions.messages.ExportChatInviteRequest(peer=entity))
+            return result.link
+        except AttributeError:
+            logger.warning("ExportChatInviteRequest not available, using alternative method")
+        except Exception as e1:
+            logger.warning(f"ExportChatInviteRequest failed: {e1}")
+        try:
+            invite_link = await client.export_chat_invite_link(entity)
+            return invite_link
+        except Exception as e2:
+            logger.warning(f"export_chat_invite_link failed: {e2}")
+        try:
+            if isinstance(entity, (Chat, Channel)):
+                full_chat = await client(functions.messages.GetFullChatRequest(chat_id=entity.id))
+                if hasattr(full_chat, "full_chat") and hasattr(full_chat.full_chat, "invite_link"):
+                    return full_chat.full_chat.invite_link or "No invite link available."
+        except Exception as e3:
+            logger.warning(f"GetFullChatRequest failed: {e3}")
+        return "Could not retrieve invite link for this chat."
+    except Exception as e:
+        logger.exception(f"get_invite_link failed (chat_id={chat_id})")
+        return log_and_format_error("get_invite_link", e, chat_id=chat_id)
+
+
+async def join_chat_by_link(link: str) -> str:
+    """
+    Join a chat by invite link.
+    """
+    try:
+        if "/" in link:
+            hash_part = link.split("/")[-1]
+            if hash_part.startswith("+"):
+                hash_part = hash_part[1:]
+        else:
+            hash_part = link
+        try:
+            invite_info = await client(functions.messages.CheckChatInviteRequest(hash=hash_part))
+            if hasattr(invite_info, "chat") and invite_info.chat:
+                chat_title = getattr(invite_info.chat, "title", "Unknown Chat")
+                return f"You are already a member of this chat: {chat_title}"
+        except Exception:
+            pass
+        result = await client(functions.messages.ImportChatInviteRequest(hash=hash_part))
+        if result and hasattr(result, "chats") and result.chats:
+            chat_title = getattr(result.chats[0], "title", "Unknown Chat")
+            return f"Successfully joined chat: {chat_title}"
+        return f"Joined chat via invite hash."
+    except Exception as e:
+        err_str = str(e).lower()
+        if "expired" in err_str:
+            return "The invite hash has expired and is no longer valid."
+        elif "invalid" in err_str:
+            return "The invite hash is invalid or malformed."
+        elif "already" in err_str and "participant" in err_str:
+            return "You are already a member of this chat."
+        logger.exception(f"join_chat_by_link failed (link={link})")
+        return f"Error joining chat: {e}"
+
+
+async def export_chat_invite(chat_id: Union[int, str]) -> str:
+    """
+    Export a chat invite link.
+    """
+    try:
+        entity = await resolve_entity(chat_id)
+        try:
+            from telethon.tl import functions
+
+            result = await client(functions.messages.ExportChatInviteRequest(peer=entity))
+            return result.link
+        except AttributeError:
+            logger.warning("ExportChatInviteRequest not available, using alternative method")
+        except Exception as e1:
+            logger.warning(f"ExportChatInviteRequest failed: {e1}")
+        try:
+            invite_link = await client.export_chat_invite_link(entity)
+            return invite_link
+        except Exception as e2:
+            logger.warning(f"export_chat_invite_link failed: {e2}")
+            return log_and_format_error("export_chat_invite", e2, chat_id=chat_id)
+    except Exception as e:
+        logger.exception(f"export_chat_invite failed (chat_id={chat_id})")
+        return log_and_format_error("export_chat_invite", e, chat_id=chat_id)
+
+
+async def import_chat_invite(hash: str) -> str:
+    """
+    Import a chat invite by hash.
+    """
+    try:
+        if hash.startswith("+"):
+            hash = hash[1:]
+        try:
+            from telethon.errors import (
+                InviteHashExpiredError,
+                InviteHashInvalidError,
+                UserAlreadyParticipantError,
+                ChatAdminRequiredError,
+                UsersTooMuchError,
+            )
+
+            invite_info = await client(functions.messages.CheckChatInviteRequest(hash=hash))
+            if hasattr(invite_info, "chat") and invite_info.chat:
+                chat_title = getattr(invite_info.chat, "title", "Unknown Chat")
+                return f"You are already a member of this chat: {chat_title}"
+        except Exception as check_err:
+            pass
+        try:
+            result = await client(functions.messages.ImportChatInviteRequest(hash=hash))
+            if result and hasattr(result, "chats") and result.chats:
+                chat_title = getattr(result.chats[0], "title", "Unknown Chat")
+                return f"Successfully joined chat: {chat_title}"
+            return f"Joined chat via invite hash."
+        except Exception as join_err:
+            err_str = str(join_err).lower()
+            if "expired" in err_str:
+                return "The invite hash has expired and is no longer valid."
+            elif "invalid" in err_str:
+                return "The invite hash is invalid or malformed."
+            elif "already" in err_str and "participant" in err_str:
+                return "You are already a member of this chat."
+            elif "admin" in err_str:
+                return "Cannot join this chat - requires admin approval."
+            elif "too much" in err_str or "too many" in err_str:
+                return "Cannot join this chat - it has reached maximum number of participants."
+            else:
+                raise
+    except Exception as e:
+        logger.exception(f"import_chat_invite failed (hash={hash})")
+        return log_and_format_error("import_chat_invite", e, hash=hash)
