+++ src/telegram_mcp/utils.py
@@ -1,0 +1,303 @@
+import os
+import time
+import asyncio
+import re
+from enum import Enum
+from functools import wraps
+from datetime import datetime
+from typing import List, Dict, Optional, Union, Any
+from pathlib import Path
+from telethon.tl.types import *
+from telethon.tl.functions.messages import GetForumTopicsRequest, ReadDiscussionRequest
+from telethon import functions, types, utils
+from mcp.server.fastmcp import Context
+
+# Use absolute imports for our own module as it's the standard for PyPI packages
+from telegram_mcp.client import client, logger
+
+
+class ValidationError(Exception):
+    """Custom exception for validation errors."""
+
+    pass
+
+
+def json_serializer(obj):
+    """Helper function to convert non-serializable objects for JSON serialization."""
+    if isinstance(obj, datetime):
+        return obj.isoformat()
+    if isinstance(obj, bytes):
+        return obj.decode("utf-8", errors="replace")
+    # Add other non-serializable types as needed
+    raise TypeError(f"Object of type {type(obj)} is not JSON serializable")
+
+
+def get_entity_type(entity: Any) -> str:
+    """Return a normalized, human-readable chat/entity type."""
+    if isinstance(entity, User):
+        return "User"
+    if isinstance(entity, Chat):
+        return "Group (Basic)"
+    if isinstance(entity, Channel):
+        if getattr(entity, "megagroup", False):
+            return "Supergroup"
+        return "Channel" if getattr(entity, "broadcast", False) else "Group"
+    return type(entity).__name__
+
+
+def get_entity_filter_type(entity: Any) -> Optional[str]:
+    """Return list_chats-compatible filter type: user/group/channel."""
+    entity_type = get_entity_type(entity)
+    if entity_type == "User":
+        return "user"
+    if entity_type in ("Group (Basic)", "Group", "Supergroup"):
+        return "group"
+    if entity_type == "Channel":
+        return "channel"
+    return None
+
+
+class ErrorCategory(str, Enum):
+    CHAT = "CHAT"
+    MSG = "MSG"
+    CONTACT = "CONTACT"
+    GROUP = "GROUP"
+    MEDIA = "MEDIA"
+    PROFILE = "PROFILE"
+    AUTH = "AUTH"
+    ADMIN = "ADMIN"
+    FOLDER = "FOLDER"
+
+
+def log_and_format_error(
+    function_name: str,
+    error: Exception,
+    prefix: Optional[Union[ErrorCategory, str]] = None,
+    user_message: str = None,
+    **kwargs,
+) -> str:
+    """
+    Centralized error handling function.
+
+    Logs an error and returns a formatted, user-friendly message.
+
+    Args:
+        function_name: Name of the function where the error occurred.
+        error: The exception that was raised.
+        prefix: Error code prefix (e.g., ErrorCategory.CHAT, "VALIDATION-001").
+            If None, it will be derived from the function_name.
+        user_message: A custom user-facing message to return. If None, a generic one is created.
+        **kwargs: Additional context parameters to include in the log.
+
+    Returns:
+        A user-friendly error message with an error code.
+    """
+    # Generate a consistent error code
+    if isinstance(prefix, str) and prefix == "VALIDATION-001":
+        # Special case for validation errors
+        error_code = prefix
+    else:
+        if prefix is None:
+            # Try to derive prefix from function name
+            for category in ErrorCategory:
+                if category.name.lower() in function_name.lower():
+                    prefix = category
+                    break
+
+        prefix_str = prefix.value if isinstance(prefix, ErrorCategory) else (prefix or "GEN")
+        error_code = f"{prefix_str}-ERR-{abs(hash(function_name)) % 1000:03d}"
+
+    # Format the additional context parameters
+    context = ", ".join(f"{k}={v}" for k, v in kwargs.items())
+
+    # Log the full technical error
+    logger.error(f"Error in {function_name} ({context}) - Code: {error_code}", exc_info=True)
+
+    # Return a user-friendly message
+    if user_message:
+        return user_message
+
+    return f"An error occurred (code: {error_code}). Check mcp_errors.log for details."
+
+
+def validate_id(*param_names_to_validate):
+    """
+    Decorator to validate chat_id and user_id parameters, including lists of IDs.
+    It checks for valid integer ranges, string representations of integers,
+    and username formats.
+    """
+
+    def decorator(func):
+        @wraps(func)
+        async def wrapper(*args, **kwargs):
+            for param_name in param_names_to_validate:
+                if param_name not in kwargs or kwargs[param_name] is None:
+                    continue
+
+                param_value = kwargs[param_name]
+
+                def validate_single_id(value, p_name):
+                    # Handle integer IDs
+                    if isinstance(value, int):
+                        if not (-(2**63) <= value <= 2**63 - 1):
+                            return (
+                                None,
+                                f"Invalid {p_name}: {value}. ID is out of the valid integer range.",
+                            )
+                        return value, None
+
+                    # Handle string IDs
+                    if isinstance(value, str):
+                        try:
+                            int_value = int(value)
+                            if not (-(2**63) <= int_value <= 2**63 - 1):
+                                return (
+                                    None,
+                                    f"Invalid {p_name}: {value}. ID is out of the valid integer range.",
+                                )
+                            return int_value, None
+                        except ValueError:
+                            if re.match(r"^@?[a-zA-Z0-9_]{5,}$", value):
+                                return value, None
+                            else:
+                                return (
+                                    None,
+                                    f"Invalid {p_name}: '{value}'. Must be a valid integer ID, or a username string.",
+                                )
+
+                    # Handle other invalid types
+                    return (
+                        None,
+                        f"Invalid {p_name}: {value}. Type must be an integer or a string.",
+                    )
+
+                if isinstance(param_value, list):
+                    validated_list = []
+                    for item in param_value:
+                        validated_item, error_msg = validate_single_id(item, param_name)
+                        if error_msg:
+                            return log_and_format_error(
+                                func.__name__,
+                                ValidationError(error_msg),
+                                prefix="VALIDATION-001",
+                                user_message=error_msg,
+                                **{param_name: param_value},
+                            )
+                        validated_list.append(validated_item)
+                    kwargs[param_name] = validated_list
+                else:
+                    validated_value, error_msg = validate_single_id(param_value, param_name)
+                    if error_msg:
+                        return log_and_format_error(
+                            func.__name__,
+                            ValidationError(error_msg),
+                            prefix="VALIDATION-001",
+                            user_message=error_msg,
+                            **{param_name: param_value},
+                        )
+                    kwargs[param_name] = validated_value
+
+            return await func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
+
+
+def format_entity(entity) -> Dict[str, Any]:
+    """Helper function to format entity information consistently."""
+    result = {"id": entity.id}
+
+    if hasattr(entity, "title"):
+        result["name"] = entity.title
+        result["type"] = "group" if isinstance(entity, Chat) else "channel"
+    elif hasattr(entity, "first_name"):
+        name_parts = []
+        if entity.first_name:
+            name_parts.append(entity.first_name)
+        if hasattr(entity, "last_name") and entity.last_name:
+            name_parts.append(entity.last_name)
+        result["name"] = " ".join(name_parts)
+        result["type"] = "user"
+        if hasattr(entity, "username") and entity.username:
+            result["username"] = entity.username
+        if hasattr(entity, "phone") and entity.phone:
+            result["phone"] = entity.phone
+
+    return result
+
+
+async def resolve_entity(identifier: Union[int, str]) -> Any:
+    """Resolve entity with automatic cache warming on miss.
+
+    StringSession has no persistent entity cache. If get_entity() fails
+    because the cache is cold (ValueError on PeerUser lookup for group IDs),
+    warm the cache via get_dialogs() and retry.
+    """
+    try:
+        return await client.get_entity(identifier)
+    except ValueError:
+        await client.get_dialogs()
+        return await client.get_entity(identifier)
+
+
+async def resolve_input_entity(identifier: Union[int, str]) -> Any:
+    """Like resolve_entity() but returns an InputPeer."""
+    try:
+        return await client.get_input_entity(identifier)
+    except ValueError:
+        await client.get_dialogs()
+        return await client.get_input_entity(identifier)
+
+
+def format_message(message) -> Dict[str, Any]:
+    """Helper function to format message information consistently."""
+    result = {
+        "id": message.id,
+        "date": message.date.isoformat(),
+        "text": message.message or "",
+    }
+
+    if message.from_id:
+        result["from_id"] = utils.get_peer_id(message.from_id)
+
+    if message.media:
+        result["has_media"] = True
+        result["media_type"] = type(message.media).__name__
+
+    return result
+
+
+def get_sender_name(message) -> str:
+    """Helper function to get sender name from a message."""
+    if not message.sender:
+        return "Unknown"
+
+    # Check for group/channel title first
+    if hasattr(message.sender, "title") and message.sender.title:
+        return message.sender.title
+    elif hasattr(message.sender, "first_name"):
+        # User sender
+        first_name = getattr(message.sender, "first_name", "") or ""
+        last_name = getattr(message.sender, "last_name", "") or ""
+        full_name = f"{first_name} {last_name}".strip()
+        return full_name if full_name else "Unknown"
+    else:
+        return "Unknown"
+
+
+def get_engagement_info(message) -> str:
+    """Helper function to get engagement metrics (views, forwards, reactions) from a message."""
+    engagement_parts = []
+    views = getattr(message, "views", None)
+    if views is not None:
+        engagement_parts.append(f"views:{views}")
+    forwards = getattr(message, "forwards", None)
+    if forwards is not None:
+        engagement_parts.append(f"forwards:{forwards}")
+    reactions = getattr(message, "reactions", None)
+    if reactions is not None:
+        results = getattr(reactions, "results", None)
+        total_reactions = sum(getattr(r, "count", 0) or 0 for r in results) if results else 0
+        engagement_parts.append(f"reactions:{total_reactions}")
+    return f" | {', '.join(engagement_parts)}" if engagement_parts else ""
